% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_available_foodnumbers}
\alias{get_available_foodnumbers}
\title{get foodnumbers from NDNS, optionally filter away those with GHGE=NA and/or cost=NA
used to set rv_foodnumbers}
\usage{
get_available_foodnumbers(ghge = F, price = F)
}
\description{
get foodnumbers from NDNS, optionally filter away those with GHGE=NA and/or cost=NA
used to set rv_foodnumbers
}
