% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_badness_per_swap}
\alias{get_badness_per_swap}
\title{this calculates for each swap WHY it occurred and what the badness_diff is.}
\usage{
get_badness_per_swap(sol, t, m)
}
\arguments{
\item{sol}{solution}

\item{t}{targets}

\item{m}{swap_from - this is base_items with the accepted_swaps added.}
}
\value{
data.frame with swaps with bad, why, good, and ori_good
}
\description{
this calculates for each swap WHY it occurred and what the badness_diff is.
}
