% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_base_items}
\alias{get_base_items}
\title{get the items (and EnergykJ amounts) from ndnse.}
\usage{
get_base_items(random_s, available_fn)
}
\arguments{
\item{random_s}{the seriali we are getting}

\item{available_fn}{vector of allowed FoodNumbers.}
}
\value{
data.frame with FoodNumber and EnergykJ
}
\description{
get the items (and EnergykJ amounts) from ndnse.
}
