% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_chosen_items}
\alias{get_chosen_items}
\title{join the accepted and the rejected foodnumbers.}
\usage{
get_chosen_items(accepted, rejected)
}
\arguments{
\item{accepted}{data.frame with FoodNumber= the accepted foodnumbers}

\item{rejected}{data.frame with FoodNumber= the rejected foodnumbers}
}
\value{
the union of the FoodNumbers from both arguments.
}
\description{
join the accepted and the rejected foodnumbers.
}
