% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_extra_items}
\alias{get_extra_items}
\title{choose  nextra items from choose_from.}
\usage{
get_extra_items(choose_from, nextra)
}
\arguments{
\item{choose_from}{vector of foodnumbers}

\item{nextra}{numbers of items to choose}
}
\value{
data.frame with FoodNumber= the chosen items; EnergykJ=0
used in r_extra_items
}
\description{
choose  nextra items from choose_from.
}
