% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_nutrient_matrix_with_badness}
\alias{get_nutrient_matrix_with_badness}
\title{lists the amounts of nutrients for sol and swap_from and shows the targets and badness for each nutrient.}
\usage{
get_nutrient_matrix_with_badness(sol, t, swap_from, mbi)
}
\arguments{
\item{sol}{solution amounts}

\item{t}{targets with directions}

\item{swap_from}{merged base items with accepted choices}

\item{mbi}{merged_base_items - the original food list.}
}
\value{
data.frame the table showing the amounts of nutrients, the targets, and the badness for sol, mbi, and swap_from
}
\description{
lists the amounts of nutrients for sol and swap_from and shows the targets and badness for each nutrient.
}
