% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_swap_from}
\alias{get_swap_from}
\title{make the swap_from list, which contains base items,}
\usage{
get_swap_from(bi, rejected, accepted, mei)
}
\value{
merged
}
\description{
make the swap_from list, which contains base items,
}
