% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{get_targets}
\alias{get_targets}
\title{calculates the number of days based on the energy in merged_base_items, then
multiplies TARGNA targets by the number of days.}
\usage{
get_targets(merged_base_items, price, price_factor, ghge, ghge_factor)
}
\arguments{
\item{merged_base_items}{data.frame with items with FoodNumber and EnergykJ}

\item{price}{logical; do we constrain on price?}

\item{price_factor}{if we constrain on price we aim to keep price under price_factor x original price}

\item{ghge}{logical; do we constrain on green house gas emissions?}

\item{ghge_factor}{if we constrain on ghge we aim to keep ghge under ghge_factor x original ghge}
}
\value{
data.frame with directions=c("<=", "<=", ">=") and targets=numeric vector.
}
\description{
calculates the number of days based on the energy in merged_base_items, then
multiplies TARGNA targets by the number of days.
}
