% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{getFoods}
\alias{getFoods}
\title{get foodnumbers with FoodNames.
used to update the FoodNumbers selectize}
\usage{
getFoods()
}
\description{
get foodnumbers with FoodNames.
used to update the FoodNumbers selectize
}
