% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{init}
\alias{init}
\title{load targets.txt from minChangeLP and load ndns_file}
\usage{
init(
  targets_file,
  nutrientbank_file,
  intake_ndns_file,
  nutrientbank_field_names
)
}
\description{
load targets.txt from minChangeLP and load ndns_file
}
