% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{make_swap_string}
\alias{make_swap_string}
\title{Using a single record from 'get_badness_per_swap' we create an english readable message about this swap.}
\usage{
make_swap_string(c)
}
\arguments{
\item{c}{a single record from 'get_badness_per_swap' - data.frame with only 1 record. contains fields new_g, ori_g, product, why, good, ori_good.}
}
\value{
data.frame the swap is described in 'swap'. the 'goodness' change is described in 'well'. the why (the value of the nutrient involved) is described in 'why'.
}
\description{
Using a single record from 'get_badness_per_swap' we create an english readable message about this swap.
}
