% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{merge_accepted}
\alias{merge_accepted}
\title{we merge new_energykJ from accepted into mbi, and override their energykJ with new_energykJ.}
\usage{
merge_accepted(mbi, accepted)
}
\description{
we merge new_energykJ from accepted into mbi, and override their energykJ with new_energykJ.
}
