% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{merge_items}
\alias{merge_items}
\title{group and sum the items by FoodNumber; then merge them with all nutrient bank columns}
\usage{
merge_items(items)
}
\arguments{
\item{items}{data.frame with FoodNumber and EnergykJ}
}
\value{
items grouped and summed by FoodNumber then merged with all nutrient bank columns. (data.frame)
}
\description{
group and sum the items by FoodNumber; then merge them with all nutrient bank columns
}
