% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodSwap.R
\name{onlyGoodFoods}
\alias{onlyGoodFoods}
\title{get foodnumbers with FoodNames.
used to update the FoodNumbers selectize}
\usage{
onlyGoodFoods()
}
\description{
get foodnumbers with FoodNames.
used to update the FoodNumbers selectize
}
